invalidate map caching
